Stop pinning `pylint`
We are many deprecation warnings in the travis logs for things that are fixed in newer versions of `pylint` or `pylint`'s dependencies.

So opening a PR to see if `master` passes on whatever version travis pulls in. Note that `pylint` >= 2.0 does not support python 2, so this will result in different versions of pylint running for python 2 vs python 3. Personally, I am just fine with this.